refactor(host): extract whisper discovery and progress helpers
Interpreter probing and tqdm parsing leave the model manager owning only
download, marker, and dispose lifecycle. Shebang reads now take the first
512 bytes instead of slurp the whole file.
